+++ packages/wrapper/web/src/pluginsStrategy/pluginsStrategy.ts
@@ -1,3 +1,4 @@
+import { FinalTheme } from './../RichContentWrapperTypes';
 import { merge } from 'lodash';
 import { RichContentProps, InitialState } from '../RichContentWrapperTypes';
 
@@ -37,7 +38,7 @@ function viewerStrategy(
   prev: ViewerPluginsStrategy,
   curr: ViewerPluginConfig,
   theme: object,
-  initialState: InitialState
+  initialState?: InitialState
 ) {
   const { type, config, typeMapper, decorator, inlineStyleMapper } = curr;
   return {
@@ -56,10 +57,11 @@ function viewerStrategy(
 export default function pluginsStrategy(
   isEditor = false,
   plugins: PluginConfig[] = [],
-  childProps: RichContentProps = {}
+  childProps: RichContentProps = {},
+  theme: FinalTheme['theme']
 ): PluginsStrategy {
   // TODO: Should consider initialState to be explicitly required in child props
-  const { theme = {}, initialState } = childProps;
+  const { initialState } = childProps;
   let strategy: EditorPluginsStrategy | ViewerPluginsStrategy;
 
   if (isEditor) {