+++ packages/wrapper/web/src/themeStrategy/themeStrategy.ts
@@ -1,9 +1,13 @@
+import { FinalTheme } from './../RichContentWrapperTypes';
 import ThemeGenerator from './ThemeGenerator';
 import jss from 'jss';
 import preset from 'jss-preset-default';
 import { defaultTheme } from './defaults';
 
-export default function themeStrategy(isEditor: boolean, themeProperties: ThemeProperties) {
+export default function themeStrategy(
+  isEditor: boolean,
+  themeProperties: ThemeProperties
+): FinalTheme {
   const { theme } = themeProperties;
   if (typeof theme === 'object') {
     return { theme: { ...defaultTheme, ...theme } };